Refactor dependencies
Persistence now depends on API and not vice versa. Enums were moved to
API layer, DTOs depends only on another DTOs, not on Entities. Mapping
fuction for Map objects has been added.
