+++ web-is-api/src/main/java/cz/muni/fi/pa165/monsterslayers/dto/CreateHeroDTO.java
@@ -1,11 +1,10 @@
 package cz.muni.fi.pa165.monsterslayers.dto;
 
-import cz.muni.fi.pa165.monsterslayers.entities.enums.PowerElement;
+import cz.muni.fi.pa165.monsterslayers.enums.PowerElement;
 import java.util.Collection;
 import java.util.HashSet;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
-import org.hibernate.validator.constraints.NotEmpty;
 /**
  * Data transfer object for creating hero to specified user.
  *
@@ -18,7 +17,8 @@ public class CreateHeroDTO {
     @Size(min = 4, max = 32)
     private String heroName;
 
-    @NotEmpty(message = "Hero should control at least one power element. Otherwise, it's not a true hero :).")
+    //TODO: we cannot have Hibernate dependency here
+//    @NotEmpty(message = "Hero should control at least one power element. Otherwise, it's not a true hero :).")
     private Collection<PowerElement> elements = new HashSet<>();
 
     public CreateHeroDTO(Long userId, String heroName) {
