+++ web-is-api/src/main/java/cz/muni/fi/pa165/monsterslayers/enums/JobStatus.java
@@ -1,8 +1,8 @@
-package cz.muni.fi.pa165.monsterslayers.entities.enums;
+package cz.muni.fi.pa165.monsterslayers.enums;
 
 /**
  * Enumeration of possible statuses of a Job
- * 
+ *
  * @author Tomáš Richter
  */
 public enum JobStatus {
