+++ web-is-api/src/main/java/cz/muni/fi/pa165/monsterslayers/facade/MonsterTypeFacade.java
@@ -3,7 +3,6 @@ package cz.muni.fi.pa165.monsterslayers.facade;
 import cz.muni.fi.pa165.monsterslayers.dto.CreateMonsterTypeDTO;
 import cz.muni.fi.pa165.monsterslayers.dto.ModifyMonsterTypeDTO;
 import cz.muni.fi.pa165.monsterslayers.dto.MonsterTypeDTO;
-import cz.muni.fi.pa165.monsterslayers.entities.MonsterType;
 
 /**
  * Facade layer interface for MonsterType.
