+++ web-is-persistence/pom.xml
@@ -17,6 +17,11 @@
         <version>1.0-SNAPSHOT</version>
     </parent>
     <dependencies>
+        <dependency>
+            <groupId>cz.muni.fi.pa165.monster-slayers</groupId>
+            <artifactId>web-is-api</artifactId>
+            <version>${project.parent.version}</version>
+        </dependency>
 
         <!-- JUnit -->
         <!--<dependency>
