+++ web-is-persistence/src/main/java/cz/muni/fi/pa165/monsterslayers/entities/Hero.java
@@ -1,6 +1,6 @@
 package cz.muni.fi.pa165.monsterslayers.entities;
 
-import cz.muni.fi.pa165.monsterslayers.entities.enums.PowerElement;
+import cz.muni.fi.pa165.monsterslayers.enums.PowerElement;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
@@ -95,7 +95,7 @@ public class Hero {
         }
         return Objects.equals(this.user, other.getUser());
     }
-    
+
     @Override
     public int hashCode() {
         int hash = 7;
