+++ web-is-service/src/main/java/cz/muni/fi/pa165/monsterslayers/service/HeroService.java
@@ -2,9 +2,6 @@ package cz.muni.fi.pa165.monsterslayers.service;
 
 import cz.muni.fi.pa165.monsterslayers.entities.Hero;
 import cz.muni.fi.pa165.monsterslayers.entities.MonsterType;
-import cz.muni.fi.pa165.monsterslayers.entities.User;
-import cz.muni.fi.pa165.monsterslayers.entities.enums.PowerElement;
-import java.util.Collection;
 import org.springframework.stereotype.Service;
 
 /**
