+++ web-is-service/src/main/java/cz/muni/fi/pa165/monsterslayers/service/MappingService.java
@@ -1,16 +1,18 @@
 package cz.muni.fi.pa165.monsterslayers.service;
-import java.util.Collection;
 import java.util.List;
+import java.util.Map;
+
 import org.springframework.stereotype.Service;
 
 /**
  * Interface for bean mapping (used for mapping entities to data transfer objects)
- * 
+ *
  * @author Tomáš Richter
  */
 
 @Service
 public interface MappingService {
     <T> List<T> mapTo(Iterable<?> objects, Class<T> mapToClass);
+    <K, V> Map<K, V> mapTo(Map<?, V> objects, Class<K> mapToClass);
     <T> T mapTo(Object u, Class<T> mapToClass);
 }
