+++ web-is-service/src/main/java/cz/muni/fi/pa165/monsterslayers/service/MappingServiceImpl.java
@@ -1,8 +1,7 @@
 package cz.muni.fi.pa165.monsterslayers.service;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
+import java.util.*;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -10,13 +9,13 @@ import org.dozer.Mapper;
 
 /**
  * Implementation of bean mapping service using Dozer framework.
- * 
+ *
  * @author Tomáš Richter
  */
 
 @Service
 public class MappingServiceImpl implements MappingService {
-    
+
     @Autowired
     private Mapper dozer;
 
@@ -27,6 +26,16 @@ public class MappingServiceImpl implements MappingService {
             mappedCollection.add(dozer.map(object, mapToClass));
         }
         return mappedCollection;
+    }
+
+    @Override
+    public <K, V> Map<K, V> mapTo(Map<?, V> objects, Class<K> mapToClass) {
+        Map<K, V> mappedMap = new HashMap<>();
+        for(Map.Entry<?, V> entry : objects.entrySet()){
+            mappedMap.put(dozer.map(entry.getKey(), mapToClass), entry.getValue());
+        }
+
+        return mappedMap;
     }
 
     @Override
